introduce `IOrderable` interface to mark content which should be "orderable" within a container — the (to be completed) adapter will then only store the order of such items, typically reducing the amount of information and its manipulation significantly...
use-cases for this are sites with only a relatively small amount of content for which (explicit) order matters.  being able to limit the "orderable" content within a container, for example to only a subset of all available types, allows to not having to separate it from "non-orderable" content.  so far you'd typically end up putting "orderable" content, for example navigation items, into regular folders and other content into "large" folders, most of the time solely for performance reasons.  this adapter will hopefully help to avoid having to make this distinction in the future and still provide the (by far) better performance of btree folders.

svn path=/plone.folder/trunk/; revision=24553